feat(cli)!: remove deprecated collect/quarantine/policy/gate categories
Deletes 4 registration-only files (src/cli/categories/{collect,
quarantine,policy,gate}.ts) which shipped 9 deprecated commands in
0.7.0. The underlying action functions (runCollectCi, runCalibrate,
runQuarantineSuggest/Apply, etc.) are retained — they back
flaker apply and other primary commands.

runCollectCi is moved from categories/collect.ts into
src/cli/commands/collect/ci.ts (its natural home) so apply.ts can
continue to import it after the category file is deleted.

BREAKING CHANGE: flaker collect*, quarantine suggest/apply,
policy quarantine/check/report, gate review/history/explain now exit
non-zero with "unknown command". See docs/migration-0.6-to-0.7.md.
